Move conversions to 'global' conversions file. Move session_id in display message to identifier_id and add identifier_type. Implement display message and session cost callbacks in OCPP 201 module.
