+++ modules/OCPP/CMakeLists.txt
@@ -29,6 +29,7 @@ target_sources(${MODULE_NAME}
         "ocpp_generic/ocppImpl.cpp"
         "session_cost/session_costImpl.cpp"
         "conversions.cpp"
+        "../../utils/conversions.cpp"
 )
 
 # ev@c55432ab-152c-45a9-9d2e-7281d50c69c3:v1
