+++ modules/OCPP201/CMakeLists.txt
@@ -29,6 +29,8 @@ target_sources(${MODULE_NAME}
         "conversions.cpp"
         "transaction_handler.cpp"
         "ocpp_generic/ocppImpl.cpp"
+        "session_cost/session_costImpl.cpp"
+        "../../utils/conversions.cpp"
 )
 
 # ev@c55432ab-152c-45a9-9d2e-7281d50c69c3:v1
