+++ modules/OCPP201/OCPP201.hpp
@@ -16,9 +16,11 @@
 #include <generated/interfaces/empty/Implementation.hpp>
 #include <generated/interfaces/ocpp/Implementation.hpp>
 #include <generated/interfaces/ocpp_data_transfer/Implementation.hpp>
+#include <generated/interfaces/session_cost/Implementation.hpp>
 
 // headers for required interface implementations
 #include <generated/interfaces/auth/Interface.hpp>
+#include <generated/interfaces/display_message/Interface.hpp>
 #include <generated/interfaces/evse_manager/Interface.hpp>
 #include <generated/interfaces/evse_security/Interface.hpp>
 #include <generated/interfaces/ocpp_data_transfer/Interface.hpp>
@@ -52,10 +54,11 @@ public:
             std::unique_ptr<auth_token_validatorImplBase> p_auth_validator,
             std::unique_ptr<auth_token_providerImplBase> p_auth_provider,
             std::unique_ptr<ocpp_data_transferImplBase> p_data_transfer, std::unique_ptr<ocppImplBase> p_ocpp_generic,
+            std::unique_ptr<session_costImplBase> p_session_cost,
             std::vector<std::unique_ptr<evse_managerIntf>> r_evse_manager, std::unique_ptr<systemIntf> r_system,
             std::unique_ptr<evse_securityIntf> r_security,
             std::vector<std::unique_ptr<ocpp_data_transferIntf>> r_data_transfer, std::unique_ptr<authIntf> r_auth,
-            Conf& config) :
+            std::vector<std::unique_ptr<display_messageIntf>> r_display_message, Conf& config) :
         ModuleBase(info),
         mqtt(mqtt_provider),
         p_main(std::move(p_main)),
@@ -63,12 +66,14 @@ public:
         p_auth_provider(std::move(p_auth_provider)),
         p_data_transfer(std::move(p_data_transfer)),
         p_ocpp_generic(std::move(p_ocpp_generic)),
+        p_session_cost(std::move(p_session_cost)),
         r_evse_manager(std::move(r_evse_manager)),
         r_system(std::move(r_system)),
         r_security(std::move(r_security)),
         r_data_transfer(std::move(r_data_transfer)),
         r_auth(std::move(r_auth)),
-        config(config){};
+        r_display_message(std::move(r_display_message)),
+        config(config) {};
 
     Everest::MqttProvider& mqtt;
     const std::unique_ptr<emptyImplBase> p_main;
@@ -76,11 +81,13 @@ public:
     const std::unique_ptr<auth_token_providerImplBase> p_auth_provider;
     const std::unique_ptr<ocpp_data_transferImplBase> p_data_transfer;
     const std::unique_ptr<ocppImplBase> p_ocpp_generic;
+    const std::unique_ptr<session_costImplBase> p_session_cost;
     const std::vector<std::unique_ptr<evse_managerIntf>> r_evse_manager;
     const std::unique_ptr<systemIntf> r_system;
     const std::unique_ptr<evse_securityIntf> r_security;
     const std::vector<std::unique_ptr<ocpp_data_transferIntf>> r_data_transfer;
     const std::unique_ptr<authIntf> r_auth;
+    const std::vector<std::unique_ptr<display_messageIntf>> r_display_message;
     const Conf& config;
 
     // ev@1fce4c5e-0ab8-41bb-90f7-14277703d2ac:v1
