+++ modules/OCPP201/manifest.yaml
@@ -48,6 +48,9 @@ provides:
   ocpp_generic:
     description: Generic OCPP interface.
     interface: ocpp
+  session_cost:
+    description: Send session cost
+    interface: session_cost
 requires:
   evse_manager:
     interface: evse_manager
@@ -68,6 +71,10 @@ requires:
   auth:
     interface: auth
     min_connections: 1
+    max_connections: 1
+  display_message:
+    interface: display_message
+    min_connections: 0
     max_connections: 1
 enable_external_mqtt: true
 enable_global_errors: true
