+++ modules/OCPP201/session_cost/session_costImpl.cpp
@@ -1,0 +1,16 @@
+// SPDX-License-Identifier: Apache-2.0
+// Copyright Pionix GmbH and Contributors to EVerest
+
+#include "session_costImpl.hpp"
+
+namespace module {
+namespace session_cost {
+
+void session_costImpl::init() {
+}
+
+void session_costImpl::ready() {
+}
+
+} // namespace session_cost
+} // namespace module
